+++ packages/app-desktop/drizzle/0004_round_bulldozer.sql
@@ -1,0 +1,4 @@
+CREATE TABLE `linked_file` (
+	`id` text PRIMARY KEY NOT NULL,
+	`filePath` text NOT NULL
+);