+++ packages/app-desktop/src/ipc/api.ts
@@ -20,6 +20,7 @@ import { WorkspacesApiBridge } from "../workspace/workspace.handler";
 import { EmbedApiBridge } from "@/embed/embed.handler";
 import { ContextMenuApiBridge } from "@/desktop-integration/context-menu.handler";
 import { ShellApiBridge } from "@/desktop-integration/shell.handler";
+import { FileLinkApiBridge } from "@/link/file-link.handler";
 
 export const DarkwriteElectronAPI = {
   note: NoteApiBridge,
@@ -51,6 +52,7 @@ export const DarkwriteElectronAPI = {
     contextMenu: ContextMenuApiBridge,
     shell: ShellApiBridge,
   },
+  fileLink: FileLinkApiBridge,
   checkUpdate: new IPCHandler(false, Updater.checkUpdate),
   backup: BackupApiBridge,
 } satisfies DarkwriteAPI;