+++ packages/common/src/index.ts
@@ -41,3 +41,4 @@ export * from "./field";
 export * from "./webutils";
 export * from "./context-menu";
 export * from "./url";
+export * from "./link";