+++ packages/frontend/src/features/editor/types.ts
@@ -32,6 +32,7 @@ export enum Block {
   Image = "dwimage",
   HorizontalRule = "horizontalRule",
   LinkToPage = "linkToPage",
+  LinkToLocalFile = "fileLink",
 }
 
 export enum TextDirection {